update check-student-against-area for the new structure
