+++ modules/web/helpers/workers/check-student-against-area/lib-log.js
@@ -1,0 +1,6 @@
+// @flow
+
+const log = (...args: any[]) =>
+    args.length && console.log('worker:check-student-against-area', ...args)
+
+export default log
